remove intValue() and change Long to Integer

+++ ewm-service/src/main/java/ru/practicum/ewm/service/event/EventRequestCount.java
@@ -2,5 +2,5 @@ package ru.practicum.ewm.service.event;
 
 public interface EventRequestCount {
     Long getEventId();
-    Long getCount();
+    Integer getCount();
 }
